対応情報の詳細画面でDBの画像が表示されない不具合の修正

+++ src/main/java/com/example/demo/shohinMaster/ShohinMasterRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ShohinMasterRepository extends JpaRepository<ShohinMaster, Long>, JpaSpecificationExecutor<ShohinMaster> {
+public interface ShohinMasterRepository extends JpaRepository<ShohinMaster, Integer>, JpaSpecificationExecutor<ShohinMaster> {
 
 	ShohinMaster findByShohinid(int shohinid);
